Update README.md
# shinros_ws
Autonomous Mobile Robot for Visitor Guidance

A mobile robot which can autonomously drive itself to reach the specified
destination. In our robot the specific destination information is given by the user
who is not familiar with the place he/she is visiting, so with the help of AMR they
can reach their desired destination. For the user interface we have created a GUI
for giving user inputs and it will give a gist of information about the place which
the user wants to go and we have changed many parameters in the ROS
Navigation stack like changing parameters in local and global planners for more precise motion of navigation, it is achieved with sensor
fusion which include wheel encoder and LIDAR data ,so it can avoid obstacles
while going to the destination. By communicating between GUI and navigation
stack in ROS, we achieve the autonomous mobile robot for visitor guidance, so
that the visitor can reach out to any places where they want to go, without any
worries about knowing each and every pathway to reach their destination.

Developed By
Ajeeth J
Barathan R
Veera Sivakumar M